+++ packages/f-vue-icons/src/components/BaseIcon/index.js
@@ -1,0 +1,7 @@
+import BaseIcon from './BaseIcon.vue';
+
+const install = Vue => {
+    Vue.component(BaseIcon.name, BaseIcon);
+};
+
+export default install;