+++ packages/f-vue-icons/src/components/index.js
@@ -1,73 +1,45 @@
 import AlertIcon from './AlertIcon';
-import AmexIcon from './AmexIcon';
-import AmexSafekeyIcon from './AmexSafekeyIcon';
+import AppStoreIcon from './AppStoreIcon';
+import BaseIcon from './BaseIcon';
 import BasketIcon from './BasketIcon';
 import CashIcon from './CashIcon';
 import ChevronIcon from './ChevronIcon';
 import ClockIcon from './ClockIcon';
 import CollectionIcon from './CollectionIcon';
 import CrossIcon from './CrossIcon';
-import DankortIcon from './DankortIcon';
 import DeliveryIcon from './DeliveryIcon';
-import FacebookIcon from './FacebookIcon';
 import InfoIcon from './InfoIcon';
-import InstagramIcon from './InstagramIcon';
-import InteracIcon from './InteracIcon';
 import LocalLegendIcon from './LocalLegendIcon';
-import MaestroIcon from './MaestroIcon';
 import MapPinIcon from './MapPinIcon';
-import MastercardIcon from './MastercardIcon';
-import MastercardSecurecodeIcon from './MastercardSecurecodeIcon';
 import MinusIcon from './MinusIcon';
 import OfferIcon from './OfferIcon';
 import PadlockIcon from './PadlockIcon';
-import PaypalIcon from './PaypalIcon';
-import PinterestIcon from './PinterestIcon';
 import PlusIcon from './PlusIcon';
-import RssIcon from './RssIcon';
 import SpicyIcon from './SpicyIcon';
 import StopwatchIcon from './StopwatchIcon';
-import TwitterIcon from './TwitterIcon';
 import VegetarianIcon from './VegetarianIcon';
-import VisaIcon from './VisaIcon';
-import VisaVerifiedIcon from './VisaVerifiedIcon';
 import WarningIcon from './WarningIcon';
-import YoutubeIcon from './YoutubeIcon';
 
 export {
     AlertIcon,
-    AmexIcon,
-    AmexSafekeyIcon,
+    AppStoreIcon,
+    BaseIcon,
     BasketIcon,
     CashIcon,
     ChevronIcon,
     ClockIcon,
     CollectionIcon,
     CrossIcon,
-    DankortIcon,
     DeliveryIcon,
-    FacebookIcon,
     InfoIcon,
-    InstagramIcon,
-    InteracIcon,
     LocalLegendIcon,
-    MaestroIcon,
     MapPinIcon,
-    MastercardIcon,
-    MastercardSecurecodeIcon,
     MinusIcon,
     OfferIcon,
     PadlockIcon,
-    PaypalIcon,
-    PinterestIcon,
     PlusIcon,
-    RssIcon,
     SpicyIcon,
     StopwatchIcon,
-    TwitterIcon,
     VegetarianIcon,
-    VisaIcon,
-    VisaVerifiedIcon,
-    WarningIcon,
-    YoutubeIcon
+    WarningIcon
 };