+++ packages/f-vue-icons/src/components/social/index.js
@@ -1,0 +1,15 @@
+import FacebookIcon from './FacebookIcon';
+import InstagramIcon from './InstagramIcon';
+import PinterestIcon from './PinterestIcon';
+import RssIcon from './RssIcon';
+import TwitterIcon from './TwitterIcon';
+import YoutubeIcon from './YoutubeIcon';
+
+export default {
+    FacebookIcon,
+    InstagramIcon,
+    PinterestIcon,
+    RssIcon,
+    TwitterIcon,
+    YoutubeIcon
+};